Move deploy function to its own file
